Reimplement CIE/FDE parsing.
Now with intelligence!  The old method tried to keep state minimal
by just declaring that a byte array *was* a CIE or FDE, and
parsing out the information requested on demand.

This was okay for the simple values at the beginning of each
entry, but became miserable beyond that.  The trouble is that all
of the structures are variable-length.  This made grabbing fields
towards the tail of the structure a pain, because it had to
read/re-parse everything before it, just so it could figure out
how long everything was and where the offset of the interesting
thing was.

Even worse, FDEs cannot be parsed without 1) knowledge of where in
the larger byte stream that the FDE starts, and 2) the CIE that is
'paired' to it.  That pairing is described by an *offset* into the
larger byte stream---thus, again, the FDE cannot be parsed alone.

This made communication very difficult in the old approach, since
it is a requirement of parsing that one bounce between the
'larger' byte stream and the specific CIE or FDE currently being
parsed.  The new implementation is much more top down: we get a
hint at what's there, and if we need more info we grab that before
passing it down to the parsing code that needs it.  This makes a
new cleaner API in terms of 'Associated' CIEs as well, since the
FDE can just include a copy of the needed CIE as a field.

+++ cfa.go
@@ -142,7 +142,7 @@ func (ixn Inst) String() string {
                      ixn.Oper[0], ixn.Oper[1])
 }
 
-func Decode(insn []byte, code_align uint, data_align int64) Inst {
+func Decode(insn []byte, code_align int, data_align int) Inst {
   rv := Inst{Op: Opcode(insn[0])}
   rv.Len = 1
   if rv.Op < CFA_advance_loc {
